cpufreq: hotplug: Update governor policy
When 2 CPUs are plugged on, use the max load to decide
increasing/decreasing the frequency. Keep on using the average
load in the CPU hotplug policy.
Update the frequency decrease policy as done by the ondemand
governor: don't wait to cross the down_threshold in order to
decrease the frequency, go directly to the lowest frequency which
can support the current CPU utilization by keeping 30% of idle in
order to not cross the up_threshold.

The down_threshold is changed to 30 instead of 20 based on experiments.

This patch shows a good performance and power impact on the different
UCs of the OMAP4 platform:
- Divide by 2 the time of the sunspider web browsing benchmark.
- Gain of ~15mW on AV record use case.

Acked-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>